feat: on page change, the image is not replaced until it has fully loaded
When the page is changed in modal, the currently visible image is not replaced by the new one until the new one has fully loaded. Images can be replaced immediately by setting the `immediateMediaReplace` option to `true` - this allows you to take advantage of progressive jpeg loading, for example, and show the progress of the image loading.