feat: `NajaContentLoader` matcher function changed
`NajaContentLoader` is now used under the same conditions as the `AjaxModalExtension` for Naja itself. Previously it was necessary to use `data-naja-modal` to trigger this content loader. But the Naja extension itself is used even without this attribute, if there is a class `ajax` on the interacted element and the modal is already opened. The `NajaContentLoader` now adds the same conditions to be consistent and easy to use. This also makes it easier to e.g. enable loading inside modal only if already opened - now you just need to add the `ajax` class to the interacted element.

The content loader also adds a data attribute `data-content-loading` to the title element when using the default loading title. This makes it easier to style the loading state when there is no relevant content yet.