fix: `PdModalNajaAdapter` properly persists width and classes
Width and class name are persisted when navigating in modal using Naja. This was previously broken when using browser history in addition to `data-naja-history="replace"`. In this case, if the first opener set width and class name and the second link inside the modal (with `data-naja-history="replace"`) didn't have these attributes, they were not stored in the History API. Now the adapter also stores these options and sets them correctly if they are not overridden by the second element.