fix: Scrolling the document when navigating inside the modal
Navigating within the modal will move the focus to the modal itself. This causes the document to scroll to the top as the browser tries to automatically scroll to the focused element. This was previously fixed when opening the modal, but not when navigating within it. Now the `scrollTop` position is handled in both situations, and the document under the modal doesn't scroll anymore.