fix: Prevent default called when handling closer click
When the content closer is clicked, the prevent default is correctly called. This allows the element to be a link, but only the modal is closed. This will allow the same element markup to be used outside of the modal (e.g. if no JS is enabled or the modal content opens in a new window).