feature: Introduced `data-modal-empty` state
When the modal with async content is opened and loading for the first time, the `data-modal-empty` attribute is set along with the `data-modal-loading`. When navigating within the modal using ajax, only `data-modal-loading` is set again, not `data-modal-empty`. This allows loading states to be distinguished, whether there is actual content or not.