fix: Refactor event handling to allow button clicks with Naja 3.2.0
Refactored the condition to call event.preventDefault() only for anchor elements, not for buttons. This change is necessary due to updates in Naja 3.2.0, where event binding no longer relies on the button’s click event but instead uses the form’s submit event. As a result, button click events must always be allowed to propagate.