fix: Correct event.preventDefault() call in autoBind
In release 2.9.0, the `event.preventDefault()` call was changed to run only when clicking an anchor element. However, the wrong variable was checked to determine whether the target is an anchor. Instead of checking `element` (`event.target`), we must check the actual `opener` element, because the clicked anchor may contain nested elements (for example `img`, `span`, etc.).